v6.0-serialize: all new features move to LanguageSpecificationV6
